feat: Add gcs interface and default implementation
This commit adds the necessary files to set up a Docker container for integration testing. The Dockerfile defines two layers, one for building the application and another for running it. The docker-compose.yml file sets up two services, one for a fake GCS server and another for the GCS client tester.

The README.md file has been updated with instructions on how to run both unit tests and integration tests using Docker or Golang.

The app package now includes an interface called BucketClient, which defines methods for interacting with a GCS bucket. The gcs package implements this interface with a struct called BucketClient that uses the cloud.google.com/go/storage library to interact with Google Cloud Storage.